requirements/common.txt
#requirements/common.txt

django==1.4

django-activelink==0.3
django-ajax-validation==0.1.3
python-dateutil==1.5

django_excel_templates==0.1
xlwt==0.7.4

django-notification==0.2

django-pipeline==1.2.6
django-smart-selects==1.0.4

markdown==2.1.1

mimeparse==0.1.3